Upgrade rake development dependency
Why you made the change:

I did this so it was ancient, as well as because there was a missmatch
between the version of rake specified and how it was used which was
having problems. Upgrading to the latest version of rake resolves this
issue so that `bundle exec rake` can successfully be run.